fix: choose countries only from list for validation
